Renames helper to protocol.
