+++ lib/processor.js
@@ -2,7 +2,7 @@ var _ = require('lodash');
 var util = require('util');
 var net = require('net');
 
-var helper = require('./helper');
+var protocol = require('./protocol');
 var filter = require('./filter');
 var projection = require('./projection');
 var update = require('./update');
@@ -34,10 +34,10 @@ that = {
     logger.trace('new socket');
     var processSocketData = function(buf) {
       var header, clientReqMsg;
-      header = helper.fromMsgHeaderBuf(buf);
+      header = protocol.fromMsgHeaderBuf(buf);
       switch (header.opCode) {
-        case helper.OP_QUERY:
-          clientReqMsg = helper.fromOpQueryBuf(header, buf);
+        case protocol.OP_QUERY:
+          clientReqMsg = protocol.fromOpQueryBuf(header, buf);
           if (clientReqMsg.fullCollectionName.match(/\.\$cmd$/)) {
             if (clientReqMsg.query) {
               if (clientReqMsg.query.findandmodify) {
@@ -53,16 +53,16 @@ that = {
             that.doQuery(socket, clientReqMsg);
           }
           break;
-        case helper.OP_INSERT:
-          clientReqMsg = helper.fromOpInsertBuf(header, buf);
+        case protocol.OP_INSERT:
+          clientReqMsg = protocol.fromOpInsertBuf(header, buf);
           that.doInsert(socket, clientReqMsg);
           break;
-        case helper.OP_DELETE:
-          clientReqMsg = helper.fromOpDeleteBuf(header, buf);
+        case protocol.OP_DELETE:
+          clientReqMsg = protocol.fromOpDeleteBuf(header, buf);
           that.doDelete(socket, clientReqMsg);
           break;
-        case helper.OP_UPDATE:
-          clientReqMsg = helper.fromOpUpdateBuf(header, buf);
+        case protocol.OP_UPDATE:
+          clientReqMsg = protocol.fromOpUpdateBuf(header, buf);
           that.doUpdate(socket, clientReqMsg);
           break;
         default:
@@ -91,7 +91,7 @@ that = {
       reply = {
         documents: { 'ismaster': true, 'ok': true }
       };
-      replyBuf = helper.toOpReplyBuf(clientReqMsg, reply);
+      replyBuf = protocol.toOpReplyBuf(clientReqMsg, reply);
       socket.write(replyBuf);
     } else if (clientReqMsg.query['getlasterror']) {
       reply = {
@@ -107,7 +107,7 @@ that = {
         logger.trace('Reporting lasterror:', that.lastError);
         delete that.lastError;
       }
-      replyBuf = helper.toOpReplyBuf(clientReqMsg, reply);
+      replyBuf = protocol.toOpReplyBuf(clientReqMsg, reply);
       socket.write(replyBuf);
     } else if (clientReqMsg.query['count']) {
       that.doCount(socket, clientReqMsg);
@@ -124,7 +124,7 @@ that = {
     var query = clientReqMsg.query.query;
     var docs = filter.filterItems(collection, query);
     var reply = {documents: {ok: true, n: docs.length}};
-    var replyBuf = helper.toOpReplyBuf(clientReqMsg, reply);
+    var replyBuf = protocol.toOpReplyBuf(clientReqMsg, reply);
     socket.write(replyBuf);
   },
 
@@ -145,7 +145,7 @@ that = {
     } catch (err) {
       if (err.message.match(/^BadValue /)) {
         var reply = {documents: [{'$err': err.message}]};
-        var replyBuf = helper.toOpReplyBuf(clientReqMsg, reply);
+        var replyBuf = protocol.toOpReplyBuf(clientReqMsg, reply);
         socket.write(replyBuf);
         return;
       } else {
@@ -160,7 +160,7 @@ that = {
     docs = docs.slice(skip, skip + limit);
     docs = projection.getProjection(docs, clientReqMsg.returnFieldSelector);
     delete that.affectedDocuments;
-    replyBuf = helper.toOpReplyBuf(clientReqMsg, { documents: docs });
+    replyBuf = protocol.toOpReplyBuf(clientReqMsg, { documents: docs });
     socket.write(replyBuf);
   },
 
@@ -237,7 +237,7 @@ that = {
   doFindAndModify: function(socket, clientReqMsg) {
     function writeErrorReply(clientReqMsg, errorMessage) {
       var reply = {documents: [{ok: false, errmsg: errorMessage}]};
-      var replyBuf = helper.toOpReplyBuf(clientReqMsg, reply);
+      var replyBuf = protocol.toOpReplyBuf(clientReqMsg, reply);
       socket.write(replyBuf);
     }
 
@@ -335,7 +335,7 @@ that = {
       [returnedDoc],
       clientReqMsg.query.fields)[0];
     var reply = {documents: {value: returnedDoc}};
-    var replyBuf = helper.toOpReplyBuf(clientReqMsg, reply);
+    var replyBuf = protocol.toOpReplyBuf(clientReqMsg, reply);
     socket.write(replyBuf);
   },
 
@@ -365,7 +365,7 @@ that = {
       });
     }
     var reply = {documents: {ok: true, values: results}};
-    var replyBuf = helper.toOpReplyBuf(clientReqMsg, reply);
+    var replyBuf = protocol.toOpReplyBuf(clientReqMsg, reply);
     socket.write(replyBuf);
   }
 };
